+++ src/PowerShell/Microsoft.WinGet.Client/Commands/Common/FinderExtendedCommand.cs
@@ -1,15 +1,12 @@
 ﻿// -----------------------------------------------------------------------------
-// <copyright file="BaseFinderExtendedCommand.cs" company="Microsoft Corporation">
+// <copyright file="FinderExtendedCommand.cs" company="Microsoft Corporation">
 //     Copyright (c) Microsoft Corporation. Licensed under the MIT License.
 // </copyright>
 // -----------------------------------------------------------------------------
 
 namespace Microsoft.WinGet.Client.Commands.Common
 {
-    using System.Collections.Generic;
     using System.Management.Automation;
-    using Microsoft.Management.Deployment;
-    using Microsoft.WinGet.Client.Attributes;
     using Microsoft.WinGet.Client.Common;
 
     /// <summary>
@@ -17,12 +14,11 @@ namespace Microsoft.WinGet.Client.Commands.Common
     /// the "search" and "list" commands. This class contains an extended set of parameters suited for
     /// that purpose.
     /// </summary>
-    public abstract class BaseFinderExtendedCommand : BaseFinderCommand
+    public abstract class FinderExtendedCommand : FinderCommand
     {
         /// <summary>
         /// Gets or sets the filter that is matched against the tags of the package.
         /// </summary>
-        [Filter(Field = PackageMatchField.Tag)]
         [Parameter(
             ParameterSetName = Constants.FoundSet,
             ValueFromPipelineByPropertyName = true)]
@@ -31,7 +27,6 @@ namespace Microsoft.WinGet.Client.Commands.Common
         /// <summary>
         /// Gets or sets the filter that is matched against the commands of the package.
         /// </summary>
-        [Filter(Field = PackageMatchField.Command)]
         [Parameter(
             ParameterSetName = Constants.FoundSet,
             ValueFromPipelineByPropertyName = true)]
@@ -45,15 +40,5 @@ namespace Microsoft.WinGet.Client.Commands.Common
             ParameterSetName = Constants.FoundSet,
             ValueFromPipelineByPropertyName = true)]
         public uint Count { get; set; }
-
-        /// <summary>
-        /// Searches for packages from configured sources.
-        /// </summary>
-        /// <param name="behavior">A <see cref="CompositeSearchBehavior" /> value.</param>
-        /// <returns>A list of <see cref="MatchResult" /> objects.</returns>
-        protected IReadOnlyList<MatchResult> FindPackages(CompositeSearchBehavior behavior)
-        {
-            return this.FindPackages(behavior, this.Count);
-        }
     }
 }
